Specify device tree changes
This gives controllers the guarantee that as long as a device is in any of the "operational" states, they can assume that the MQTT structure of that device won't change.
This also implies that each time a device goes back online, all nodes and properties must be re-checked for any changes.