Add implementation details and other things (#200)
* Explicitly note that all values must be sent during initialization

I stumbled over this in a comment in one of the discussions. The specification is pretty vague about this at the moment. So I added a sentence that makes this requirement more explicit and less hard to miss.

* Specify device tree changes

This gives controllers the guarantee that as long as a device is in any of the "operational" states, they can assume that the MQTT structure of that device won't change.
This also implies that each time a device goes back online, all nodes and properties must be re-checked for any changes.

* Add implementation notes

* Replace "QoS 1" with "At least once (1)"

In future, when referring to QoS levels, use:

- "At most once (0)"
- "At least once (1)"
- "Exactly once (2)"

Closes #196

* Make events (non-retained properties) use a QoS of 2

Closes #166

* Update convention.md

spelling fix

* Revert "Make events (non-retained properties) use a QoS of 2"

This reverts commit 4a6ea00ee27c1a52a7419912029b4fb9e050ec99.

* Make events (non-retained properties) use a QoS of 0

* Improve implementation notes

Co-authored-by: Michael Cumming <mike@4831.com>